Modifications of metrics, split model/litmodel, new OSE configuration running
