+++ .gitignore
@@ -8,3 +8,9 @@ tb_profile
 *out
 submit_4DVarNet.slurm
 log
+spde.py
+scipy_sparse_tools.py
+models_wSPDE.py
+merge_ose_osse*
+main_wSPDE.slurm
+checkpoints
