+++ config_q/__init__.py
@@ -1,8 +1,9 @@
 # Specify the dataset spatial bounds
 dim_range = {
-    'lat': slice(33, 43),
-    'lon': slice(-65, -55),
-
+    #'lat': slice(33, 43),
+    #'lon': slice(-65, -55),
+    'lat': slice(27, 57),
+    'lon': slice(-77, 3),
 }
 # Specify the batch patch size
 slice_win = {
@@ -17,6 +18,11 @@ strides = {
     'lon': 200,
 }
 
+time_period = {
+    'train_slices' : (slice('2012-10-01', "2012-11-20"), slice('2013-02-07', "2013-09-30")),
+    'test_slices' : (slice('2013-01-03', "2013-01-27"),),
+    'val_slices' : (slice('2012-11-30', "2012-12-24"),),
+}
 
 params = {
     'files_cfg' : dict(
@@ -57,8 +63,8 @@ params = {
 
     # NN architectures and optimization parameters
     'batch_size'      : 2, #16#4#4#8#12#8#256#
-    'DimAE'           : 50, #10#10#50
-    'dim_grad_solver' : 150,
+    'DimAE'           : 25, #10#10#50
+    'dim_grad_solver' : 70,
     'dropout'         : 0.25,
     'dropout_phi_r'   : 0.,
