+++ config_q/nad_swot_anim.py
@@ -1,0 +1,11 @@
+from . import params
+
+params['files_cfg'].update( 
+    dict(
+        obs_mask_path='/gpfsstore/rech/yrf/commun/NATL60/NATL/data_new/dataset_nadir_0d_swot.nc',
+        obs_mask_var='ssh_mod',
+    )
+)
+
+params['animate'] = True
+
